requirements.txt
torch
transformers
chromadb
openai
tqdm
colorama
python-dotenv 